fix: use fixed-point arithmetic for bid/ask price calculation

+++ src/main/java/com/replicatorfx/model/PairState.java
@@ -7,12 +7,12 @@ import java.util.concurrent.atomic.AtomicReference;
 public final class PairState {
 
     public final AtomicReference<PairConfig> config;
-    public volatile double currentMid;
-    public volatile long   lastTickNano;
+    public volatile long currentMid;    // fixed-point, scaled by 10^decimalPlaces
+    public volatile long lastTickNano;
 
     public PairState(PairConfig config) {
-        this.config      = new AtomicReference<>(config);
-        this.currentMid  = config.initialMidPrice;
+        this.config       = new AtomicReference<>(config);
+        this.currentMid   = config.initialMidPrice;
         this.lastTickNano = 0L;
     }
 }
